refact(api): merge analytics_models.py into schemas_analytics.py (#5996)
Eliminates dedicated model file in favour of domain schema file.
Updated 3 importers: analytics_code.py, analytics_controller.py, analytics.py.
