+++ autobot-backend/api/analytics_code.py
@@ -14,7 +14,7 @@ from datetime import datetime, timezone
 
 from fastapi import APIRouter, Depends, HTTPException
 
-from api.analytics_models import CodeAnalysisRequest
+from api.schemas_analytics import CodeAnalysisRequest
 from auth_middleware import check_admin_permission
 from autobot_shared.error_boundaries import ErrorCategory, with_error_handling
 from autobot_shared.security.path_validator import validate_path
